+++ miniapp/src/main/java/com/rakuten/tech/mobile/miniapp/errors/MiniAppDownloadFileError.kt
@@ -1,15 +1,19 @@
 package com.rakuten.tech.mobile.miniapp.errors
 
+import androidx.annotation.Keep
+
 /**
  * A class to provide the custom errors specific for file download.
  */
-class MiniAppDownloadFileError(val type: String? = null, val message: String? = null) :
+@Keep
+class MiniAppDownloadFileError(val type: String? = null, val message: String? = null, val code: Int? = null) :
     MiniAppBridgeError(type, message) {
 
     companion object {
         private const val DownloadFailedError = "DownloadFailedError"
         private const val InvalidUrlError = "InvalidUrlError"
         private const val SaveFailureError = "SaveFailureError"
+        private const val DownloadHttpError = "DownloadHttpError"
 
         // Failed to download file.
         val downloadFailedError = MiniAppDownloadFileError(DownloadFailedError, errorDescription(DownloadFailedError))
@@ -19,12 +23,11 @@ class MiniAppDownloadFileError(val type: String? = null, val message: String? = 
 
         val saveFailureError = MiniAppDownloadFileError(SaveFailureError, errorDescription(SaveFailureError))
 
-        /**
-         *  send custom error message from host app.
-         *  @property code error code send to miniapp.
-         *  @property reason error message send to mini app.
-         */
-        fun custom(code: String, reason: String) = MiniAppDownloadFileError(code, reason)
+        internal fun httpError(code: Int, message: String) = MiniAppDownloadFileError(
+            type = DownloadHttpError,
+            code = code,
+            message = message
+        )
 
         private fun errorDescription(error: String): String {
             return when (error) {