+++ miniapp/src/main/java/com/rakuten/tech/mobile/miniapp/js/MessageBridgeRatDispatcher.kt
@@ -19,6 +19,8 @@ internal class MessageBridgeRatDispatcher(private val miniAppAnalytics: MiniAppA
     internal fun getAcType(action: String): Actype {
         return when (action) {
             ActionType.GET_UNIQUE_ID.action -> Actype.GET_UNIQUE_ID
+            ActionType.GET_MESSAGING_UNIQUE_ID.action -> Actype.GET_MESSAGING_UNIQUE_ID
+            ActionType.GET_MAUID.action -> Actype.GET_MAUID
             ActionType.REQUEST_PERMISSION.action -> Actype.REQUEST_PERMISSION
             ActionType.REQUEST_CUSTOM_PERMISSIONS.action -> Actype.REQUEST_CUSTOM_PERMISSIONS
             ActionType.SHARE_INFO.action -> Actype.SHARE_INFO
@@ -34,6 +36,12 @@ internal class MessageBridgeRatDispatcher(private val miniAppAnalytics: MiniAppA
             ActionType.SEND_MESSAGE_TO_CONTACT_ID.action -> Actype.SEND_MESSAGE_TO_CONTACT_ID
             ActionType.SEND_MESSAGE_TO_MULTIPLE_CONTACTS.action -> Actype.SEND_MESSAGE_TO_MULTIPLE_CONTACTS
             ActionType.GET_HOST_ENVIRONMENT_INFO.action -> Actype.GET_HOST_ENVIRONMENT_INFO
+            ActionType.FILE_DOWNLOAD.action -> Actype.FILE_DOWNLOAD
+            ActionType.SECURE_STORAGE_SET_ITEMS.action -> Actype.SECURE_STORAGE_SET_ITEMS
+            ActionType.SECURE_STORAGE_GET_ITEM.action -> Actype.SECURE_STORAGE_GET_ITEM
+            ActionType.SECURE_STORAGE_REMOVE_ITEMS.action -> Actype.SECURE_STORAGE_REMOVE_ITEMS
+            ActionType.SECURE_STORAGE_CLEAR.action -> Actype.SECURE_STORAGE_CLEAR
+            ActionType.SECURE_STORAGE_SIZE.action -> Actype.SECURE_STORAGE_SIZE
             else -> Actype.DEFAULT
         }
     }