+++ miniapp/src/test/java/com/rakuten/tech/mobile/miniapp/js/hostenvironment/HostEnvironmentInfoBridgeSpec.kt
@@ -25,15 +25,18 @@ class HostEnvironmentInfoBridgeSpec : BridgeCommon() {
 
     @Before
     fun setup() {
-        When calling miniAppBridge.createBridgeExecutor(webViewListener) itReturns bridgeExecutor
-        miniAppBridge.init(
-                activity = TestActivity(),
+        ActivityScenario.launch(TestActivity::class.java).onActivity { activity ->
+            When calling miniAppBridge.createBridgeExecutor(webViewListener) itReturns bridgeExecutor
+            miniAppBridge.init(
+                activity = activity,
                 webViewListener = webViewListener,
                 customPermissionCache = mock(),
                 downloadedManifestCache = mock(),
                 miniAppId = TEST_MA_ID,
-                ratDispatcher = mock()
-        )
+                ratDispatcher = mock(),
+                secureStorageDispatcher = mock()
+            )
+        }
     }
 
     private val hostEnvCallbackObj = CallbackObj(
@@ -60,7 +63,8 @@ class HostEnvironmentInfoBridgeSpec : BridgeCommon() {
                     customPermissionCache = mock(),
                     downloadedManifestCache = mock(),
                     miniAppId = TEST_MA_ID,
-                    ratDispatcher = mock()
+                    ratDispatcher = mock(),
+                    secureStorageDispatcher = mock()
             )
             miniAppBridge.postMessage(hostEnvJsonStr)
             verify(bridgeExecutor).postValue(hostEnvCallbackObj.id, Gson().toJson(hostEnvironmentInfo))
@@ -69,22 +73,26 @@ class HostEnvironmentInfoBridgeSpec : BridgeCommon() {
 
     @Test
     fun `postError should be called when cannot retrieve environment info`() {
-        val miniAppBridge = Mockito.spy(createMiniAppMessageBridge(false))
-        val infoErrMessage = "{\"type\":\"{\\\"type\\\":\\\"Cannot get host environment info: error_message\\\"}\"}"
-        val webViewListener = createErrorWebViewListener(infoErrMessage)
-        val bridgeExecutor = Mockito.spy(miniAppBridge.createBridgeExecutor(webViewListener))
-        When calling miniAppBridge.createBridgeExecutor(webViewListener) itReturns bridgeExecutor
-        miniAppBridge.init(
-                activity = TestActivity(),
+        ActivityScenario.launch(TestActivity::class.java).onActivity { activity ->
+            val miniAppBridge = Mockito.spy(createMiniAppMessageBridge(false))
+            val infoErrMessage =
+                "{\"type\":\"{\\\"type\\\":\\\"Cannot get host environment info: error_message\\\"}\"}"
+            val webViewListener = createErrorWebViewListener(infoErrMessage)
+            val bridgeExecutor = Mockito.spy(miniAppBridge.createBridgeExecutor(webViewListener))
+            When calling miniAppBridge.createBridgeExecutor(webViewListener) itReturns bridgeExecutor
+            miniAppBridge.init(
+                activity = activity,
                 webViewListener = webViewListener,
                 customPermissionCache = mock(),
                 downloadedManifestCache = mock(),
                 miniAppId = TEST_MA_ID,
-                ratDispatcher = mock()
-        )
-        miniAppBridge.postMessage(hostEnvJsonStr)
+                ratDispatcher = mock(),
+                secureStorageDispatcher = mock()
+            )
+            miniAppBridge.postMessage(hostEnvJsonStr)
 
-        verify(bridgeExecutor).postError(TEST_CALLBACK_ID, infoErrMessage)
+            verify(bridgeExecutor).postError(TEST_CALLBACK_ID, infoErrMessage)
+        }
     }
 
     @Test