+++ testapp/src/main/java/com/rakuten/tech/mobile/testapp/ui/permission/MiniAppDownloadedListActivity.kt
@@ -47,8 +47,6 @@ class MiniAppDownloadedListActivity(private val miniApp: MiniApp) : BaseActivity
         executeLoadingList()
     }
 
-
-
     override fun onCreate(savedInstanceState: Bundle?) {
         super.onCreate(savedInstanceState)
         showBackIcon()